Disable shared help system, always use proxied help
A shared help system makes little sense in a multi-user environment,
even though it may be convenient in a single-user setting.
Reducing the number of ways in which the help system works
also makes the code simpler and thus more easy to maintain.
The proxy system is needed anyways for the dataTable() with AJAX
and for View().