Added initial support for FileUpload widgets. These do not work yet.
For some reasons, FileUpload widgets do not work. They show up in the frontend.
But clicking the button and specifying a file does not lead to a message being
sent to the kernel. This seems to be an issue with the frontend. However, the
documentation does not help me (yet) in getting an idea why this does not work
and what requirements are not met.