Assume ipywidgets version 8.0
Since there is no reliable way to enquire the correct version from
the frontend (not that I did not try), the version of ipywidgets package, which
includes the frontend Javascript code, is assumed to be 8.0.0 by default,
unless it is specified by setting the appropriate option,
e.g. `options(widget_version=c(4,5,2))`. If option
"widgets_version_from_jupyter" is set to TRUE, then the attempt to deducing the
right version numbers is made by calling `juypter --version`.