Added support for special comments '#+opt:' and '#+par:'
With a special comment starting with '#+opt:' it is now possible to
set the options locally for a single notebook cell. The same could be achieved
by calling 'RKernel::cell.options()' but this will lead to errors if
the code of the cell is exported into a script or Rmarkdown file, etc.

In the same way, a special comment starting with '#+par:' can be used
to set the graphics parameters for a single notebook cell.