Fix preprocessing of code cells.
The R-function `deparse()` does not work as I expected when there are
linebreaks in string literatals. This lead to spurious output if
the code contained string literals with linebreaks.
This problem does not arise if 'parse' is used instead of `str2expression()`
and `as.character()` is applied to the "srcref" attribute of the result
of `parse()`.

This hopefully fixes issue #7